+++ doc/source/conf.py
@@ -26,6 +26,7 @@ extensions = [
     "sphinx.ext.autodoc",
     "sphinx.ext.autosummary",
     "sphinx.ext.napoleon",
+    "sphinx.ext.autosectionlabel",
     "sphinx_exec_code",
     "sphinx_copybutton",
     "execinject",
@@ -34,6 +35,11 @@ extensions = [
 templates_path = ["_templates"]
 exclude_patterns = []
 
+autosectionlabel_prefix_document = True
+
+suppress_warnings = [
+    "autosectionlabel",
+]
 
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output