+++ CHANGELOG.md
@@ -5,7 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.12.0] - 2024-11-??
+## [0.13.0] - 2025-01-??
+
+### fixed
+
+- cli: issue with `list` command on inherited profiles
+
+### added
+
+- launchers : `.system` launcher new field `command_as_str`
+- launchers : `.system` launcher new field `expand_first_arg`
+
+### chores
+
+- Better launchers and launcher-plugin documentation.
+
+
+## [0.12.0] - 2024-11-15
 
 ### added